fix test #433 + clean up #26

+++ src/test/beast/evolution/operator/UniformIntegerOperatorTest.java
@@ -71,7 +71,7 @@ public class UniformIntegerOperatorTest extends TestCase {
 	            uniformOperator.proposal();
 	            Integer [] values = parameter.getValues();
 	            for (int k = 0; k < values.length; k++) {
-	                int j = (int)values[k];
+	                int j = values[k];
 	                count[k][j] += 1; 
 	            }
 	        }
